using_google_api/requirements.txt
chromadb==0.4.24
google-generativeai==0.5.3
langchain==0.2.12
langchain-community==0.2.5
langchain-google-genai==1.0.4
sentence-transformers==2.6.1
python-dotenv==1.0.1
tiktoken==0.7.0
